+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] -
+
+### Added
+
+- Add `wait()` reactive keyword: `wait(twin;reactive=true)`.
+
+- Add `when` keyword to `query_*` payload.
+
+### Changed
+
+- Set `archiver_interval` default to 1 second.
+
+### Fixed
+
+- Restore topics spaces at startup
+
 ## [1.2.2] - 2026-01-03
 
 ### Changed